requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-generativeai==0.7.2
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
anyio[trio]>=4.0
numpy>=1.26
